+++ pyproject.toml
@@ -3,10 +3,10 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project.scripts]
-shapeaxiflo = "src.saxi_folds:cml"
+shapeaxi = "src.saxi_folds:cml"
 
 [project]
-name = "shapeaxiflo"
+name = "shapeaxi"
 dynamic = ["version"]
 description = 'Shape Analysis Exploration and Interpretability'
 readme = "README.md"
@@ -27,7 +27,7 @@ classifiers = [
   "Programming Language :: Python :: Implementation :: PyPy",
   "Environment :: GPU :: NVIDIA CUDA :: 11.7",
 ]
-
+  
 dependencies = [
   "frozenlist==1.3.3",
   "seaborn==0.13.0",
