+++ src/saxi_logger.py
@@ -2,15 +2,17 @@ from pytorch_lightning.callbacks import Callback
 import torchvision
 import torch
 
+# This file saxi_train uses SaxiImageLogger which is a callback intended for logging images to the PyTorch Lightning logger during training
 
 class SaxiImageLogger(Callback):
+    # Periodically log visualizations of input images during training based on the value of log_steps
     def __init__(self, num_images=12, log_steps=10):
         self.log_steps = log_steps
         self.num_images = num_images
     def on_train_batch_end(self, trainer, pl_module, outputs, batch, batch_idx):        
-
+        # This function is called at the end of each training batch
         if batch_idx % self.log_steps == 0:
-
+                
                 V, F, CN, Y = batch
 
                 batch_size = V.shape[0]
@@ -21,7 +23,7 @@ class SaxiImageLogger(Callback):
                 CN = CN.to(pl_module.device, non_blocking=True).to(torch.float32)
 
                 with torch.no_grad():
-
+                    # Render the input surface mesh to an image
                     X, PF = pl_module.render(V[0:1], F[0:1], CN[0:1])
                     
                     grid_X = torchvision.utils.make_grid(X[0, 0:num_images, 0:3, :, :])#Grab the first image, RGB channels only, X, Y. The time dimension is on dim=1
@@ -32,13 +34,14 @@ class SaxiImageLogger(Callback):
 
 
 class SaxiImageLoggerNeptune(Callback):
+    # This callback logs images for visualization during training, with the ability to log images to the Neptune logging system for easy monitoring and analysis
     def __init__(self, num_images=12, log_steps=10):
         self.log_steps = log_steps
         self.num_images = num_images
-    def on_train_batch_end(self, trainer, pl_module, outputs, batch, batch_idx): 
-        
-        if batch_idx % self.log_steps == 0:
 
+    def on_train_batch_end(self, trainer, pl_module, outputs, batch, batch_idx): 
+        # This function is called at the end of each training batch
+        if batch_idx % self.log_steps == 0:
 
             V, F, CN, Y = batch
 
@@ -50,7 +53,7 @@ class SaxiImageLoggerNeptune(Callback):
             CN = CN.to(pl_module.device, non_blocking=True).to(torch.float32)
 
             with torch.no_grad():
-
+                # Render the input surface mesh to an image
                 X, PF = pl_module.render(V[0:1], F[0:1], CN[0:1])
                 
                 grid_X = torchvision.utils.make_grid(X[0, 0:num_images, 0:3, :, :])#Grab the first image, RGB channels only, X, Y. The time dimension is on dim=1
@@ -68,20 +71,24 @@ class SaxiImageLoggerNeptune(Callback):
                 trainer.logger.experiment["images/x_depth"].upload(fig)
                 plt.close()
 
+# The difference is the backend used for image logging and visualization.
+# SaxiImageLogger logs images directly within the PyTorch Lightning logging system.
+# SaxiImageLoggerNeptune sends images to the Neptune platform for more advanced experiment tracking and visualization. 
 
-###########################----TEETH----###########################
+#########################################################################SEGMENTATION PART#########################################################################################
 
 
 class TeethNetImageLogger(Callback):
+    # Another custom callback used especially for the Segmentation part for logging and visualizing images during training within a PyTorch Lightning-based deep learning workflow
     def __init__(self, num_images=12, log_steps=10):
         self.log_steps = log_steps
         self.num_images = num_images
+        
     def on_train_batch_end(self, trainer, pl_module, outputs, batch, batch_idx):        
 
         if batch_idx % self.log_steps == 0:
 
                 V, F, YF, CN = batch
-
                 batch_size = V.shape[0]
                 num_images = min(batch_size, self.num_images)
 
@@ -91,7 +98,7 @@ class TeethNetImageLogger(Callback):
                 CN = CN.to(pl_module.device, non_blocking=True).to(torch.float32)
 
                 with torch.no_grad():
-
+                    # Render the input surface mesh to an image
                     x, X, PF = pl_module((V[0:1], F[0:1], CN[0:1]))
 
                     y = torch.take(YF, PF).to(torch.int64)*(PF >= 0) # YF=input, pix_to_face=index. shape of y = shape of pix_to_face
