+++ src/saxi_nets.py
@@ -9,7 +9,7 @@ from torchvision import models
 from torchvision import transforms
 import torchmetrics
 
-import src.utils as utils
+import utils
 
 import monai
 from pytorch3d.renderer import (
@@ -22,6 +22,7 @@ from pytorch3d.structures import Meshes
 import pytorch_lightning as pl
 
 class ProjectionHead(nn.Module):
+    # Projection MLP
     def __init__(self, input_dim=1280, hidden_dim=1280, output_dim=128):
         super().__init__()
         self.output_dim = output_dim
@@ -40,6 +41,7 @@ class ProjectionHead(nn.Module):
         return x, x_v
 
 class TimeDistributed(nn.Module):
+    # Wrapper to apply a module to each time step of a sequence
     def __init__(self, module):
         super(TimeDistributed, self).__init__()
         self.module = module
@@ -67,31 +69,25 @@ class TimeDistributed(nn.Module):
         return output
 
 class SelfAttention(nn.Module):
+    # Self attention layer
     def __init__(self, in_units, out_units):
-        #super(SelfAttention, self).__init__()
         super().__init__()
-
-
         self.W1 = nn.Linear(in_units, out_units)
         self.V = nn.Linear(out_units, 1)
 
     def forward(self, query, values):        
-
-        # score shape == (batch_size, max_length, 1)
-        # we get 1 at the last axis because we are applying score to self.V
-        # the shape of the tensor before applying self.V is (batch_size, max_length, units)
-
         score = nn.Sigmoid()(self.V(nn.Tanh()(self.W1(query))))
-        
         attention_weights = score/torch.sum(score, dim=1,keepdim=True)
-
-        # context_vector shape after sum == (batch_size, hidden_size)
         context_vector = attention_weights * values
         context_vector = torch.sum(context_vector, dim=1)
-
         return context_vector, score
 
+
+
+#################################################### SEGMENTATION PART ####################################################
+
 class SaxiSegmentation(pl.LightningModule):
+    # Saxi segmentation network
     def __init__(self, args = None, out_channels=3, class_weights=None, image_size=320, radius=1.35, subdivision_level=1, train_sphere_samples=4):
 
         super(SaxiSegmentation, self).__init__()        
@@ -107,14 +103,7 @@ class SaxiSegmentation(pl.LightningModule):
         self.loss = monai.losses.DiceCELoss(include_background=False, to_onehot_y=True, softmax=True, ce_weight=self.class_weights)
         self.accuracy = torchmetrics.Accuracy(task='multiclass', num_classes=self.out_channels)
 
-        unet = monai.networks.nets.UNet(
-            spatial_dims=2,
-            in_channels=4,   # images: torch.cuda.FloatTensor[batch_size,224,224,4]
-            out_channels=out_channels, 
-            channels=(16, 32, 64, 128, 256),
-            strides=(2, 2, 2, 2),
-            num_res_units=2,
-        )
+        unet = monai.networks.nets.UNet(spatial_dims=2,in_channels=4,out_channels=out_channels, channels=(16, 32, 64, 128, 256),strides=(2, 2, 2, 2),num_res_units=2,)
         self.model = TimeDistributed(unet)
 
         ico_verts, ico_faces, ico_edges = utils.PolyDataToTensors(utils.CreateIcosahedron(radius=radius, sl=subdivision_level))
@@ -128,65 +117,48 @@ class SaxiSegmentation(pl.LightningModule):
         self.register_buffer("ico_verts", ico_verts)
 
         cameras = FoVPerspectiveCameras()
-        raster_settings = RasterizationSettings(
-            image_size=image_size, 
-            blur_radius=0, 
-            faces_per_pixel=1,
-            max_faces_per_bin=200000
-        )        
-        rasterizer = MeshRasterizer(
-            cameras=cameras, 
-            raster_settings=raster_settings
-        )
+        raster_settings = RasterizationSettings(image_size=image_size, blur_radius=0, faces_per_pixel=1, max_faces_per_bin=200000)        
+        rasterizer = MeshRasterizer(cameras=cameras, raster_settings=raster_settings)
+
         lights = AmbientLights()
-        self.renderer = MeshRenderer(
-                rasterizer=rasterizer,
-                shader=HardPhongShader(cameras=cameras, lights=lights)
-        )
+        self.renderer = MeshRenderer(rasterizer=rasterizer, shader=HardPhongShader(cameras=cameras, lights=lights))
 
     def configure_optimizers(self):
+        # Configure the optimizer
         optimizer = torch.optim.AdamW(self.parameters(), lr=self.args.lr)
         return optimizer
 
     def to(self, device=None):
+        # Move the renderer to the specified device
         self.renderer = self.renderer.to(device)
         return super().to(device)
 
     def forward(self, x):
-
+        # Forward pass
         V, F, CN = x
-        
         X, PF = self.render(V, F, CN)
         x = self.model(X)
         
         return x, X, PF
 
     def render(self, V, F, CN):
-
+        # Render the input surface mesh to an image
         textures = TexturesVertex(verts_features=CN)
         meshes = Meshes(verts=V, faces=F, textures=textures)        
-
         X = []
         PF = []
 
         for camera_position in self.ico_verts:
-
             camera_position = camera_position.unsqueeze(0)
-
             R = look_at_rotation(camera_position, device=self.device)  # (1, 3, 3)
             T = -torch.bmm(R.transpose(1, 2), camera_position[:,:,None])[:, :, 0]   # (1, 3)
-
             images = self.renderer(meshes_world=meshes.clone(), R=R, T=T)
-        
             fragments = self.renderer.rasterizer(meshes.clone())
             pix_to_face = fragments.pix_to_face
             zbuf = fragments.zbuf
-
             images = torch.cat([images[:,:,:,0:3], zbuf], dim=-1)
             images = images.permute(0,3,1,2)
-
             pix_to_face = pix_to_face.permute(0,3,1,2)
-            
             X.append(images.unsqueeze(1))
             PF.append(pix_to_face.unsqueeze(1))
         
@@ -196,23 +168,17 @@ class SaxiSegmentation(pl.LightningModule):
         return X, PF
 
     def training_step(self, train_batch, batch_idx):
-
+        # Training step
         V, F, YF, CN = train_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)
         YF = YF.to(self.device, non_blocking=True)
         CN = CN.to(self.device, non_blocking=True).to(torch.float32)
-
         x, X, PF = self((V, F, CN))
-
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
         y = y.permute(0, 2, 1, 3, 4)
-            
         loss = self.loss(x, y)
-
         batch_size = V.shape[0]
         self.log('train_loss', loss, batch_size=batch_size)
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
@@ -221,58 +187,50 @@ class SaxiSegmentation(pl.LightningModule):
         return loss
 
 
-
     def validation_step(self, val_batch, batch_idx):
+        # Validation step
         V, F, YF, CN = val_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)
         YF = YF.to(self.device, non_blocking=True)
         CN = CN.to(self.device, non_blocking=True).to(torch.float32)
-
         x, X, PF = self((V, F, CN))
-
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
         y = y.permute(0, 2, 1, 3, 4)
-            
         loss = self.loss(x, y)
-        
         batch_size = V.shape[0]
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
         self.log("val_acc", self.accuracy, batch_size=batch_size, sync_dist=True)
         self.log('val_loss', loss, batch_size=batch_size, sync_dist=True)
 
+
     def test_step(self, batch, batch_idx):
-
+        # Test step
         V, F, YF, CN = val_batch
-
         x, X, PF = self(V, F, CN)
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-        
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, h, w -> batch, channels, time, h, w
         y = y.permute(0, 2, 1, 3, 4) 
-
         loss = self.loss(x, y)
-
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))        
-
         return {'test_loss': loss, 'test_correct': self.accuracy}
 
-class SaxiClassification(pl.LightningModule):
-    def __init__(self, **kwargs):
 
+
+#################################################### CLASSIFICATION PART ####################################################
+
+class SaxiClassification(pl.LightningModule):
+    # Saxi classification network
+    def __init__(self, **kwargs):
         super(SaxiClassification, self).__init__()
         self.save_hyperparameters()
-
         self.class_weights = None
+
         if hasattr(self.hparams, 'class_weights'):
             self.class_weights = torch.tensor(self.hparams.class_weights).to(torch.float32)            
-        
         self.loss = nn.CrossEntropyLoss(weight=self.class_weights)
         self.accuracy = torchmetrics.Accuracy(task='multiclass', num_classes=self.hparams.out_classes)
-        
 
         if hasattr(monai.networks.nets, self.hparams.base_encoder):
             template_model = getattr(monai.networks.nets, self.hparams.base_encoder)
@@ -282,9 +240,7 @@ class SaxiClassification(pl.LightningModule):
             raise "{base_encoder} not in monai networks or torchvision".format(base_encoder=self.hparams.base_encoder)
 
         model_params = eval('dict(%s)' % self.hparams.base_encoder_params.replace(' ',''))
-        
         self.convnet = template_model(**model_params)
-
         self.F = TimeDistributed(self.convnet)
         self.V = nn.Linear(self.hparams.hidden_dim, self.hparams.hidden_dim)
         self.A = SelfAttention(in_units=self.hparams.hidden_dim, out_units=64)
@@ -292,26 +248,15 @@ class SaxiClassification(pl.LightningModule):
 
         cameras = FoVPerspectiveCameras()
 
-        raster_settings = RasterizationSettings(
-            image_size=self.hparams.image_size, 
-            blur_radius=0, 
-            faces_per_pixel=1,
-            max_faces_per_bin=200000
-        )        
-        rasterizer = MeshRasterizer(
-            cameras=cameras, 
-            raster_settings=raster_settings
-        )
+        raster_settings = RasterizationSettings(image_size=self.hparams.image_size, blur_radius=0, faces_per_pixel=1,max_faces_per_bin=200000)        
+        rasterizer = MeshRasterizer(cameras=cameras, raster_settings=raster_settings)
         lights = AmbientLights()
-        self.renderer = MeshRenderer(
-                rasterizer=rasterizer,
-                shader=HardPhongShader(cameras=cameras, lights=lights)
-        )
-
+        self.renderer = MeshRenderer(rasterizer=rasterizer,shader=HardPhongShader(cameras=cameras, lights=lights))
         self.ico_sphere(radius=self.hparams.radius, subdivision_level=self.hparams.subdivision_level)
 
 
     def ico_sphere(self, radius=1.35, subdivision_level=1):
+        # Create an icosphere
         ico_verts, ico_faces, ico_edges = utils.PolyDataToTensors(utils.CreateIcosahedron(radius=radius, sl=subdivision_level))
         ico_verts = ico_verts.to(torch.float32)
 
@@ -321,15 +266,21 @@ class SaxiClassification(pl.LightningModule):
         
         self.register_buffer("ico_verts", ico_verts)
 
+
     def configure_optimizers(self):
+        # Configure the optimizer
         optimizer = torch.optim.AdamW(self.parameters(), lr=self.hparams.lr)
         return optimizer
 
+
     def to(self, device=None):
+        # Move the renderer to the specified device
         self.renderer = self.renderer.to(device)
         return super().to(device)
 
+
     def forward(self, X):
+        # Forward pass
         x_f = self.F(X)
         x_v = self.V(x_f)
         x_a, x_s = self.A(x_f, x_v)
@@ -337,32 +288,25 @@ class SaxiClassification(pl.LightningModule):
         
         return x, x_s
 
+
     def render(self, V, F, CN):
-        
+        # Render the input surface mesh to an image
         textures = TexturesVertex(verts_features=CN.to(torch.float32))
         meshes = Meshes(verts=V, faces=F, textures=textures)
-
         X = []
         PF = []
 
         for camera_position in self.ico_verts:
-
             camera_position = camera_position.unsqueeze(0)
-
             R = look_at_rotation(camera_position, device=self.device)  # (1, 3, 3)
             T = -torch.bmm(R.transpose(1, 2), camera_position[:,:,None])[:, :, 0]   # (1, 3)
-
             images = self.renderer(meshes_world=meshes.clone(), R=R, T=T)        
-            
             fragments = self.renderer.rasterizer(meshes.clone())
             pix_to_face = fragments.pix_to_face
             zbuf = fragments.zbuf
-
             images = torch.cat([images[:,:,:,0:3], zbuf], dim=-1)
             images = images.permute(0,3,1,2)
-
             pix_to_face = pix_to_face.permute(0,3,1,2)
-            
             X.append(images.unsqueeze(1))
             PF.append(pix_to_face.unsqueeze(1))
         
@@ -372,19 +316,14 @@ class SaxiClassification(pl.LightningModule):
         return X, PF
 
     def training_step(self, train_batch, batch_idx):
-
+        # Training step
         V, F, CN, Y = train_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)        
         CN = CN.to(self.device, non_blocking=True)
-
         X, PF = self.render(V, F, CN)
-
         x, _ = self(X)
-            
         loss = self.loss(x, Y)
-
         batch_size = V.shape[0]
         self.log('train_loss', loss, batch_size=batch_size)
         self.accuracy(x, Y)
@@ -394,41 +333,27 @@ class SaxiClassification(pl.LightningModule):
 
 
     def validation_step(self, val_batch, batch_idx):
-
+        # Validation step
         V, F, CN, Y = val_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)        
         CN = CN.to(self.device, non_blocking=True)
-
         X, PF = self.render(V, F, CN)
         x, _ = self(X)
-            
         loss = self.loss(x, Y)
-
         batch_size = V.shape[0]
         self.log('val_loss', loss, batch_size=batch_size, sync_dist=True)
         self.accuracy(x, Y)
         self.log("val_acc", self.accuracy, batch_size=batch_size, sync_dist=True)
 
-    # def test_step(self, test_batch, batch_idx):
 
-    #     V, F, CN, Y = test_batch
 
-    #     V = V.to(self.device, non_blocking=True)
-    #     F = F.to(self.device, non_blocking=True)        
-    #     CN = CN.to(self.device, non_blocking=True)
 
-    #     X, PF = self.render(V, F, CN)
-    #     x, _ = self(X)
+#################################################### SAXIREGRESSION PART ####################################################
 
-    #     batch_size = V.shape[0]
-    #     self.accuracy(x, Y)
-    #     self.log("test_acc", self.accuracy, batch_size=batch_size, sync_dist=True)
-
-    #     return 1.0
 
 class SaxiRegression(pl.LightningModule):
+    # Saxi regression network
     def __init__(self, **kwargs):
 
         super(SaxiRegression, self).__init__()
@@ -474,6 +399,7 @@ class SaxiRegression(pl.LightningModule):
 
 
     def ico_sphere(self, radius=1.35, subdivision_level=1):
+        # Create an icosphere
         ico_verts, ico_faces, ico_edges = utils.PolyDataToTensors(utils.CreateIcosahedron(radius=radius, sl=subdivision_level))
         ico_verts = ico_verts.to(torch.float32)
 
@@ -483,15 +409,21 @@ class SaxiRegression(pl.LightningModule):
         
         self.register_buffer("ico_verts", ico_verts)
 
+
     def configure_optimizers(self):
+        # Configure the optimizer
         optimizer = torch.optim.AdamW(self.parameters(), lr=self.hparams.lr)
         return optimizer
 
+
     def to(self, device=None):
+        # Move the renderer to the specified device
         self.renderer = self.renderer.to(device)
         return super().to(device)
 
+
     def forward(self, X):
+        # Forward pass
         x_f = self.F(X)
         x_v = self.V(x_f)
         x_a, x_s = self.A(x_f, x_v)
@@ -499,54 +431,41 @@ class SaxiRegression(pl.LightningModule):
         
         return x
 
+
     def render(self, V, F, CN):
-        
+        # Render the input surface mesh to an image
         textures = TexturesVertex(verts_features=CN.to(torch.float32))
         meshes = Meshes(verts=V, faces=F, textures=textures)
-
         X = []
         PF = []
 
         for camera_position in self.ico_verts:
-
             camera_position = camera_position.unsqueeze(0)
-
             R = look_at_rotation(camera_position, device=self.device)  # (1, 3, 3)
             T = -torch.bmm(R.transpose(1, 2), camera_position[:,:,None])[:, :, 0]   # (1, 3)
-
             images = self.renderer(meshes_world=meshes.clone(), R=R, T=T)        
-            
             fragments = self.renderer.rasterizer(meshes.clone())
             pix_to_face = fragments.pix_to_face
             zbuf = fragments.zbuf
-
             images = torch.cat([images[:,:,:,0:3], zbuf], dim=-1)
             images = images.permute(0,3,1,2)
-
             pix_to_face = pix_to_face.permute(0,3,1,2)
-            
             X.append(images.unsqueeze(1))
             PF.append(pix_to_face.unsqueeze(1))
-        
         X = torch.cat(X, dim=1)
         PF = torch.cat(PF, dim=1)        
 
         return X, PF
 
     def training_step(self, train_batch, batch_idx):
-
+        # Training step
         V, F, CN, Y = train_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)        
         CN = CN.to(self.device, non_blocking=True)
-
         X, PF = self.render(V, F, CN)
-
         x = self(X)
-            
         loss = self.loss(x, Y)
-
         batch_size = V.shape[0]
         self.log('train_loss', loss, batch_size=batch_size)
 
@@ -554,25 +473,23 @@ class SaxiRegression(pl.LightningModule):
 
 
     def validation_step(self, val_batch, batch_idx):
-
+        # Validation step
         V, F, CN, Y = val_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)        
         CN = CN.to(self.device, non_blocking=True)
-
         X, PF = self.render(V, F, CN)
         x = self(X)
-            
         loss = self.loss(x, Y)
-
         batch_size = V.shape[0]
         self.log('val_loss', loss, batch_size=batch_size, sync_dist=True)
 
 
 
+#################################################### MONAIUNET PART ####################################################
 
 class MonaiUNet(pl.LightningModule):
+    # Monai UNet network
     def __init__(self, args = None, out_channels=3, class_weights=None, image_size=320, radius=1.35, subdivision_level=1, train_sphere_samples=4):
 
         super(MonaiUNet, self).__init__()        
@@ -626,14 +543,17 @@ class MonaiUNet(pl.LightningModule):
         )
 
     def configure_optimizers(self):
+        # Configure the optimizer
         optimizer = torch.optim.AdamW(self.parameters(), lr=self.args.lr)
         return optimizer
 
     def to(self, device=None):
+        # Move the renderer to the specified device
         self.renderer = self.renderer.to(device)
         return super().to(device)
 
     def forward(self, x):
+        # Forward pass
         V, F, CN = x
         X, PF = self.render(V, F, CN)
         x = self.model(X)
@@ -641,31 +561,23 @@ class MonaiUNet(pl.LightningModule):
 
 
     def render(self, V, F, CN):
-
+        # Render the input surface mesh to an image
         textures = TexturesVertex(verts_features=CN)
         meshes = Meshes(verts=V, faces=F, textures=textures)        
-
         X = []
         PF = []
 
         for camera_position in self.ico_verts:
-
             camera_position = camera_position.unsqueeze(0)
-
             R = look_at_rotation(camera_position, device=self.device)  # (1, 3, 3)
             T = -torch.bmm(R.transpose(1, 2), camera_position[:,:,None])[:, :, 0]   # (1, 3)
-
             images = self.renderer(meshes_world=meshes.clone(), R=R, T=T)
-        
             fragments = self.renderer.rasterizer(meshes.clone())
             pix_to_face = fragments.pix_to_face
             zbuf = fragments.zbuf
-
             images = torch.cat([images[:,:,:,0:3], zbuf], dim=-1)
             images = images.permute(0,3,1,2)
-
             pix_to_face = pix_to_face.permute(0,3,1,2)
-            
             X.append(images.unsqueeze(1))
             PF.append(pix_to_face.unsqueeze(1))
         
@@ -675,23 +587,17 @@ class MonaiUNet(pl.LightningModule):
         return X, PF
 
     def training_step(self, train_batch, batch_idx):
-
+        # Training step
         V, F, YF, CN = train_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)
         YF = YF.to(self.device, non_blocking=True)
         CN = CN.to(self.device, non_blocking=True).to(torch.float32)
-
         x, X, PF = self((V, F, CN))
-
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
         y = y.permute(0, 2, 1, 3, 4)
-            
         loss = self.loss(x, y)
-
         batch_size = V.shape[0]
         self.log('train_loss', loss, batch_size=batch_size)
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
@@ -700,41 +606,30 @@ class MonaiUNet(pl.LightningModule):
         return loss
 
 
-
     def validation_step(self, val_batch, batch_idx):
+        # Validation step   
         V, F, YF, CN = val_batch
-
         V = V.to(self.device, non_blocking=True)
         F = F.to(self.device, non_blocking=True)
         YF = YF.to(self.device, non_blocking=True)
         CN = CN.to(self.device, non_blocking=True).to(torch.float32)
-
         x, X, PF = self((V, F, CN))
-
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
         y = y.permute(0, 2, 1, 3, 4)
-            
         loss = self.loss(x, y)
-        
         batch_size = V.shape[0]
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
         self.log("val_acc", self.accuracy, batch_size=batch_size, sync_dist=True)
         self.log('val_loss', loss, batch_size=batch_size, sync_dist=True)
 
+
     def test_step(self, val_batch, batch_idx):
-
         V, F, YF, CN = val_batch
-
         x, X, PF = self((V, F, CN))
         y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
-        
         x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, h, w -> batch, channels, time, h, w
         y = y.permute(0, 2, 1, 3, 4) 
-
         loss = self.loss(x, y)
-
         self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))        
-
         return {'test_loss': loss, 'test_correct': self.accuracy}
