+++ src/saxi_predict.py
@@ -9,21 +9,23 @@ import torch
 from torch import nn
 from torch.utils.data import DataLoader
 
-import src.saxi_nets as saxi_nets
-from src.saxi_dataset import SaxiDataset, RandomRemoveTeethTransform, UnitSurfTransform
-from src.saxi_transforms import EvalTransform
+import saxi_nets 
+from saxi_dataset import SaxiDataset
+from saxi_transforms import EvalTransform, RandomRemoveTeethTransform, UnitSurfTransform, UnitSurfTransform2
 
 import pickle
 from tqdm import tqdm
 
-import src.utils as utils
+import utils
 from vtk.util.numpy_support import vtk_to_numpy, numpy_to_vtk
 import nrrd
 
+# This file proposes a prediction with the test data. It calls SaxiDataset which is a custom class from PyTorch that inherits from torch.utils.data.Datset.
+# It calls also EvalTransform
 
 
 def main(args):
-
+    # Read of the test data from a CSV or Parquet file
     mount_point = args.mount_point
 
     fname = os.path.basename(args.csv)    
@@ -35,67 +37,49 @@ def main(args):
         df = pd.read_parquet(args.csv)
     
     if args.nn == "SaxiClassification" or args.nn == "SaxiRegression":
-        
+        # The dataset and corresponding data loader are initialized for evaluation purposes.
         SAXINETS = getattr(saxi_nets, args.nn)
         model = SAXINETS.load_from_checkpoint(args.model)
-
         model.ico_sphere(args.radius, args.subdivision_level)
-
         model.to(torch.device('cuda:0'))
         model.eval()
-
         scale_factor = None
-
         if model.hparams.scale_factor:
             scale_factor = model.hparams.scale_factor
-
         test_ds = SaxiDataset(df, transform=EvalTransform(scale_factor), **vars(args))
         test_loader = DataLoader(test_ds, batch_size=1, num_workers=args.num_workers, pin_memory=True)
 
     elif args.nn == "SaxiSegmentation":
-
+        # The dataset and corresponding data loader are initialized for evaluation purposes.
         class_weights = None
         out_channels = 34
-
-        # model = saxi_nets.MonaiUNet(args, out_channels = out_channels, class_weights=class_weights, image_size=320, subdivision_level=2)
-        # model.model.module.load_state_dict(torch.load(args.model)
         MONAI = getattr(saxi_nets, args.nn)
         model = MONAI.load_from_checkpoint(args.model)
-
-
-        ds = SaxiDataset(df, mount_point = args.mount_point, transform=UnitSurfTransform(), surf_column="surf")
-
+        ds = SaxiDataset(df, mount_point = args.mount_point, transform=UnitSurfTransform2(), surf_column="surf")
         dataloader = DataLoader(ds, batch_size=1, num_workers=args.num_workers, persistent_workers=True, pin_memory=True)
-        
-
         device = torch.device('cuda')
         model.to(device)
         model.eval()
-
         softmax = torch.nn.Softmax(dim=2)
 
     with torch.no_grad():
 
         if args.nn == "SaxiClassification":
-
+            # The prediction is performed on the test data
             probs = []
             predictions = []
             softmax = nn.Softmax(dim=1)
 
             for idx, (V, F, CN, L) in tqdm(enumerate(test_loader), total=len(test_loader)):
-                
+                # The generated CAM is processed and added to the input surface mesh (surf) as a point data array
                 V = V.cuda(non_blocking=True)
                 F = F.cuda(non_blocking=True)
                 CN = CN.cuda(non_blocking=True)
-
                 X, PF = model.render(V, F, CN)
                 x, x_s = model(X)
-                
                 x = softmax(x).detach()
-                
                 probs.append(x)
                 predictions.append(torch.argmax(x, dim=1, keepdim=True))
-
 
             probs = torch.cat(probs).detach().cpu().numpy()
             predictions = torch.cat(predictions).cpu().numpy().squeeze()
@@ -103,7 +87,6 @@ def main(args):
             out_dir = os.path.join(args.out, os.path.basename(args.model))
             if not os.path.exists(out_dir):
                 os.makedirs(out_dir)
-
 
             out_probs = os.path.join(out_dir, fname.replace(ext, "_probs.pickle"))
             pickle.dump(probs, open(out_probs, 'wb'))
@@ -118,23 +101,20 @@ def main(args):
 
 
         elif args.nn == "SaxiRegression":        
-            
+            # The prediction is performed on the test data
             predictions = []
             softmax = nn.Softmax(dim=1)
 
             for idx, (V, F, CN, L) in tqdm(enumerate(test_loader), total=len(test_loader)):
-                
+                # The generated CAM is processed and added to the input surface mesh (surf) as a point data array
                 V = V.cuda(non_blocking=True)
                 F = F.cuda(non_blocking=True)
                 CN = CN.cuda(non_blocking=True)
-
                 X, PF = model.render(V, F, CN)
                 x, x_s = model(X)
-
                 predictions.append(x.detach())
             
             predictions = torch.cat(predictions).cpu().numpy().squeeze()
-
             out_dir = os.path.join(args.out, os.path.basename(args.model))
             if not os.path.exists(out_dir):
                 os.makedirs(out_dir)
@@ -149,21 +129,17 @@ def main(args):
     
 
         elif args.nn == "SaxiSegmentation":
-
+            # The prediction is performed on the test data
             for idx, batch in tqdm(enumerate(dataloader), total=len(dataloader)):
-
+                # The generated CAM is processed and added to the input surface mesh (surf) as a point data array
                 V, F, CN = batch
-
                 V = V.cuda(non_blocking=True)
                 F = F.cuda(non_blocking=True)
                 CN = CN.cuda(non_blocking=True).to(torch.float32)
-
                 x, X, PF = model((V, F, CN))
                 x = softmax(x*(PF>=0))
-
                 P_faces = torch.zeros(out_channels, F.shape[1]).to(device)
                 V_labels_prediction = torch.zeros(V.shape[1]).to(device).to(torch.int64)
-
                 PF = PF.squeeze()
                 x = x.squeeze()
 
@@ -171,20 +147,16 @@ def main(args):
                     P_faces[:, pf] += pred
 
                 P_faces = torch.argmax(P_faces, dim=0)
-
                 faces_pid0 = F[0,:,0]
                 V_labels_prediction[faces_pid0] = P_faces
-
                 surf = ds.getSurf(idx)
-
                 V_labels_prediction = numpy_to_vtk(V_labels_prediction.cpu().numpy())
                 V_labels_prediction.SetName(args.array_name)
                 surf.GetPointData().AddArray(V_labels_prediction)
-
                 output_fn = os.path.join(args.out, df["surf"][idx])
                 output_fn = output_fn.replace("./", "")
-                
                 out_dir = os.path.dirname(output_fn)
+
                 if not os.path.exists(out_dir):
                     os.makedirs(out_dir)
 
@@ -193,20 +165,18 @@ def main(args):
             out_dir = os.path.join(args.out, os.path.basename(args.model))
             if not os.path.exists(out_dir):
                 os.makedirs(out_dir)
-
             data = {
                 "surf": df["surf"],
                 "pred": [os.path.join(args.out, df["surf"][idx]) for idx in range(len(df))]
             }
-            
             result_df = pd.DataFrame(data)
-
             output_fn = os.path.join(out_dir, fname.replace(ext, "_prediction.csv"))
             print(f"Saving results to {output_fn}")
             result_df.to_csv(output_fn, index=False)              
 
 
 def get_argparse():
+    # This function defines the arguments for the prediction
     parser = argparse.ArgumentParser(description='Saxi prediction')    
 
     model_group = parser.add_argument_group('Trained')
